Some light reorganization of classes
