+++ src/main/scala/scalaz/parsers/tc/applicativeerror.scala
@@ -1,4 +1,4 @@
-package scalaz.parsers
+package scalaz.parsers.tc
 
 import scalaz.tc._
