+++ src/main/scala/scalaz/parsers/tc/productfunctor.scala
@@ -1,4 +1,6 @@
-package scalaz.parsers
+package scalaz.parsers.tc
+
+import scalaz.parsers./\
 
 trait ProductFunctorClass[F[_]] {
   def and[A, B](fa: F[A], fb: F[B]): F[A /\ B]
